+++ src/AudioOut.hpp
@@ -1,5 +1,5 @@
 // AudioOut.hpp
 #ifndef AUDIO_H
 #define AUDIO_H
-void initAudio(int *wave_table);
+void initAudio(float* wave_table);
 #endif
